requirements.txt
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.28.1
pandas==2.1.3
plotly==5.17.0
supabase==2.0.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dateutil==2.8.2 